src/third_party/blink/web_tests/external/wpt/css/css-text/parsing/letter-spacing-computed-expected.txt

This is a testharness.js-based test.
FAIL Property letter-spacing value 'normal' computes to '0px' assert_equals: expected "0px" but got "normal"
PASS Property letter-spacing value '10px' computes to '10px'
PASS Property letter-spacing value '-20px' computes to '-20px'
PASS Property letter-spacing value 'calc(10px - 0.5em)' computes to '-10px'
Harness: the test ran to completion.
